config_schema.js
In this commit, great changes and corrections to the bugs of the configuration schema script were done to make it better and less prone to issues. The patterns for `org`, `app`, `function Name`, and `stage` properties were adjusted to enforce the naming convention rules and avoid problems connected with incorrect input handling. The script has been enhanced to have better pattern matching and validation when it comes to strings to ensure that any input given is in the right format and is in conformity with the expected pattern to avoid as many errors as possible during runtime. 
 
 Furthermore, the current schema for functions and stages was enhanced to accommodate even advanced architectures especially in terms of observability and plugin management. This makes the configurations to be more relaxed and dynamic especially when handling specific properties of a provider or custom plugins. We also introduced better management of deprecated features where we highlighted such features in the schema to make sure that the users are aware of the possibility of outdated configurations. 
 
 In addition, the error management systems have been improved to give better feedback to users in the event of validation failure, thus aiding configuration debugging. The `config Validation Mode` was modified to support various modes including `error`, `warn`, and `off` to enable the user to choose how strict the validation should be. It also sets the stage for the next releases when the strict validation will be the default one for this script. 
 
 All together these improvements enhance the stability of the configuration schema, ease of use and conformity to modern software development practices and thus, prepare the ground for better deployment and configuration management.